+++ tests/functional/test_add.py
@@ -27,7 +27,7 @@ def test_add_sends_encrypted_secret(cli, gpg, api, users):
     resource = EncryptedResourceFactory(gpg=gpg, recipient=users[0].username)
     api.endpoints['add_resource'] = to_foreign_resource_response(resource)
 
-    run_add_command(cli, get_add_resource_inputs(resource) + ('',), resource.secret)
+    run_add_command(cli, get_add_resource_inputs(resource) + ('', ''), resource.secret)
 
     assert api.mocks['add_resource'].called
     assert gpg.decrypt(api.mocks['add_resource'].call_args[0][-1]['secrets'][0]['data']) == resource.secret
@@ -37,7 +37,7 @@ def test_tags_are_set(cli, gpg, api, users):
     resource = EncryptedResourceFactory(gpg=gpg, recipient=users[0].username, tags=('foo', 'bar', '#pub'))
     api.endpoints['add_resource'] = to_foreign_resource_response(resource)
 
-    run_add_command(cli, get_add_resource_inputs(resource) + ('',), resource.secret)
+    run_add_command(cli, get_add_resource_inputs(resource) + ('', ''), resource.secret)
 
     assert_tags_added(api, resource, resource.tags)
 
@@ -46,7 +46,7 @@ def test_empty_tags_does_not_call_add_tags(cli, gpg, api, users):
     resource = EncryptedResourceFactory(gpg=gpg, recipient=users[0].username, tags=[])
     api.endpoints['add_resource'] = to_foreign_resource_response(resource)
 
-    run_add_command(cli, get_add_resource_inputs(resource) + ('',), resource.secret)
+    run_add_command(cli, get_add_resource_inputs(resource) + ('', ''), resource.secret)
 
     assert not api.mocks['add_tags'].called
 
@@ -55,32 +55,32 @@ def test_add_with_sharing_encrypts_data_for_recipient(cli, gpg, api, users):
     resource = EncryptedResourceFactory(gpg=gpg, recipient=users[0].username)
     api.endpoints['add_resource'] = to_foreign_resource_response(resource)
 
-    run_add_command(cli, get_add_resource_inputs(resource) + (users[1].username,), resource.secret)
+    run_add_command(cli, get_add_resource_inputs(resource) + (users[1].username, ''), resource.secret)
 
     assert_resource_shared(api, resource, [users[1]], gpg)
 
 
 def test_added_resource_is_shared_with_default_recipients(cli, gpg, api, users):
     config = copy.deepcopy(default_config)
-    config['sharing']['default_recipients'] = 'All, {}'.format(users[1].username)
+    config['sharing']['default_owners'] = 'All, {}'.format(users[1].username)
     resource = EncryptedResourceFactory(gpg=gpg, recipient=users[0].username)
     api.endpoints['add_resource'] = to_foreign_resource_response(resource)
 
-    run_add_command(cli, get_add_resource_inputs(resource) + (users[1].username,), resource.secret, config=config)
+    run_add_command(cli, get_add_resource_inputs(resource) + (users[1].username, ''), resource.secret, config=config)
 
     assert_resource_shared(api, resource, users, gpg)
 
 
 def test_invalid_default_recipient_shows_error(cli, gpg, api, users):
     config = copy.deepcopy(default_config)
-    config['sharing']['default_recipients'] = 'inexistent@localhost'
+    config['sharing']['default_owners'] = 'inexistent@localhost'
     resource = EncryptedResourceFactory(gpg=gpg, recipient=users[0].username)
     api.add_resource = to_foreign_resource_response(resource)
 
-    result = run_add_command(cli, get_add_resource_inputs(resource) + (users[1].username,), resource.secret,
+    result = run_add_command(cli, get_add_resource_inputs(resource) + (users[1].username, ''), resource.secret,
                              config=config)
 
-    assert 'Invalid default recipient' in result.output
+    assert 'Invalid recipient' in result.output
     assert not api.mocks['share_resource'].called
 
 
@@ -88,7 +88,7 @@ def test_group_recipient_shares_with_all_group_members(cli, gpg, api, users):
     resource = EncryptedResourceFactory(gpg=gpg, recipient=users[0].username)
     api.endpoints['add_resource'] = to_foreign_resource_response(resource)
 
-    run_add_command(cli, get_add_resource_inputs(resource) + ('All',), resource.secret)
+    run_add_command(cli, get_add_resource_inputs(resource) + ('All', ''), resource.secret)
 
     assert_resource_shared(api, resource, users, gpg)
 
@@ -97,6 +97,7 @@ def test_share_with_invalid_recipient_shows_error(cli, gpg, api, users):
     resource = EncryptedResourceFactory(gpg=gpg, recipient=users[0].username)
     api.endpoints['add_resource'] = to_foreign_resource_response(resource)
 
-    result = run_add_command(cli, get_add_resource_inputs(resource) + ('foobar', users[1].username), resource.secret)
+    result = run_add_command(cli, get_add_resource_inputs(resource) + ('foobar', users[1].username, ''),
+                             resource.secret)
 
     assert 'Recipient foobar is invalid' in result.output